REVERTME [IOTG]: Android: Bypass some checking to allow hdcp to run
Android will need a refine way of checking if the
daemon is already running. In Android hdcp directory
was created in initrc script with different mode and
ownership to the path. This was a legacy approach
prior hdcpd is opensource.
Also workaround to include daemon init call. This init
is currently not working on Android.

Revert: revert the patch once the daemoin init error is root cause
and fix.

Change-Id: Ic86d3d9b55e20054792052721430159a23c3a886
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-70074
Signed-off-by: Romli, Khairul Anuar <khairul.anuar.romli@intel.com>
Reviewed-on: https://android.intel.com:443/649486

Change-Id: If8ca4c25fc7425373e4ad111b0bb208698e99ccf
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-70532
Signed-off-by: Romli, Khairul Anuar <khairul.anuar.romli@intel.com>
Reviewed-on: https://android.intel.com:443/649828